% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{bench_cal}
\alias{bench_cal}
\title{Apply bench calibrations}
\usage{
bench_cal(prh)
}
\arguments{
\item{prh}{A PRH object. Should already be decimated (see
\code{\link{decimate_prh}}) and trimmed (see \code{\link{trim_prh}})}
}
\value{
A PRH object with updated rawA, rawG, rawM, At, Gt, and Mt slots.
}
\description{
\code{bench_cal} applies bench calibrations to raw data. Calibrates inertial
sensors and fills in the At, Gt, and Mt slots. At, Mt, and Gt are in units of
g, microteslas, and radians per second, respectively, in the tag's frame.
Then calibrates the pressure sensor and fills in the p field in the data
slot. Also fills in the t and camon fields in the data slot.
}
