% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\name{bench_calib}
\alias{bench_calib}
\title{Apply bench calibrations}
\usage{
bench_calib(raw_data, acc_cal, mag_cal_on, mag_cal_off, gyr_cal)
}
\arguments{
\item{raw_data}{A tibble of the raw data (e.g. returned from \code{decimate})}

\item{acc_cal, mag_cal_on, mag_cal_off, gyr_cal}{A 2x3 matrix with the
calibration slopes and constants in the (first and second) rows and the x,
y, and z values in the columns.}
}
\value{
A list of three nx3 matrices with the calibrated records.
}
\description{
\code{bench_calib} applies bench calibrations to the raw accelerometer,
magnetometer, and gyroscope data. Outputs will be in units of g, microtesla,
and radians per second, respectively. For efficiency's sake, should only be
applied to decimated data.
}
