% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\name{decimate}
\alias{decimate}
\title{Decimate raw data}
\usage{
decimate(raw_data, new_freq)
}
\arguments{
\item{raw_data}{A tibble of the raw data (e.g. returned from
\code{import_cats})}

\item{new_freq}{The new sampling frequency. Must be a factor of the old
sampling frequency.}
}
\value{
A tibble with the same fields as the input.
}
\description{
\code{decimate} returns a subset of the raw data at a lower sampling
frequency.
}
\details{
Decimation chooses every nth record, so the new sampling frequency must be a
factor of the original. E.g. 400 Hz can be decimated to 10 Hz but not to 12
Hz.
}
\seealso{
\link{\code{import_cats}}
}
