% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\name{decimate_prh}
\alias{decimate_prh}
\title{Decimate raw data}
\usage{
decimate_prh(prh, new_freq)
}
\arguments{
\item{prh}{A PRH object.}

\item{new_freq}{The new sampling frequency. Must be a factor of the old
sampling frequency.}
}
\value{
A PRH with the freq slot filled and rawdata decimated to the new
  frequency.
}
\description{
\code{decimate_prh} returns a subset of the raw data at a lower sampling
frequency.
}
\details{
Decimation chooses every nth record, so the new sampling frequency must be a
factor of the original. E.g. 400 Hz can be decimated to 10 Hz but not to 12
Hz.
}
\seealso{
\code{\link{import_cats}}
}
