% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_rawdata.R
\name{fix_backskip}
\alias{fix_backskip}
\title{Fix backwards timestamp skips}
\usage{
fix_backskip(prh, thr = 2)
}
\arguments{
\item{prh}{A PRH object.}

\item{thr}{A numeric defining the duration threshold (in sec) for warning
about out-of-order phases.}
}
\value{
A PRH object with the rawdata slot updated
}
\description{
\code{fix_backskip} attempts to fix out-of-order timesteps.
}
\details{
Sometimes CATS timesteps skip forwards and then backwards, which is trickier
than a simple gap. \code{fix_backskip} uses a running diff of the timestamps
to look for phases with a large positive diff followed by a negative diff.
These skips are then removed with a warning if longer than a threshold (2s by
default).
}
\examples{
\dontrun{
# There's an example from 2018-06-07 15:49:46.46 UTC to 2018-06-07
# 15:49:49.352 UTC in deployment mn180607-44
# Issues a warning there's out-of-order timestamps
prh <- init_prh(PRH("mn180607-44", "44"))
# Decimation fails
prh_deci <- decimate_prh(prh)
# Fix it!
prh_fix <- fix_backskip(prh)
# Now decimation works
prh_deci <- decimate_prh(prh_fix, 10)
}
}
\seealso{
\code{\link{import_cats}}
}
