% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_rawdata.R
\name{fix_gap}
\alias{fix_gap}
\title{Fix timestamp gaps}
\usage{
fix_gap(prh, thr = 2)
}
\arguments{
\item{prh}{A PRH object.}

\item{thr}{A numeric defining the duration threshold (in sec) for warning
about out-of-order phases (2 s by default).}
}
\value{
A PRH object with the rawdata slot updated
}
\description{
\code{fix_gap} attempts to fix gaps in timesteps. Fix backwards time skips
(\code{\link{fix_backskip}}) before filling gaps.
}
\examples{
\dontrun{
# There's an example at 2018-06-07 17:31:03.04 UTC in deployment mn180607-44
# Issues a warning there's out-of-order timestamps
prh <- init_prh(PRH("mn180607-44", "44"))
# Decimation fails
prh_deci <- decimate_prh(prh)
# Fix it!
prh_fix <- fix_backskip(prh)
# Now decimation works
prh_deci <- decimate_prh(prh_fix, 10)
}
}
\seealso{
\code{\link{import_cats}}, \link{\code{fix_backskip}}
}
