% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_cats_path}
\alias{get_cats_path}
\title{Get path to CATS folder}
\usage{
get_cats_path()
}
\value{
a string with the path to the CATS folder
}
\description{
\code{get_cats_path} gets the file path to the CATS folder from the config
file.
}
