% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_access.R
\name{get_tag_path}
\alias{get_tag_path}
\title{Get the file path to tag data}
\usage{
get_tag_path(tagid, raw = FALSE)
}
\arguments{
\item{tagid}{A string with the tagid (e.g. "mn180105-22a")}

\item{raw}{A logical indicating whether to get the path to the raw or
processed tag data. FALSE by default.}
}
\value{
A string with the file path to the tag data.
}
\description{
\code{get_tag_path} gets the path to tag data (raw or processed) in the CATS
folder.
}
\details{
The COPYCATS drive should have all the processed data but only the most
recent raw data. If \code{get_tag_path} is throwing an error unexpectedly
double check that the raw data you're looking for is on the hard drive.
}
