% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_access.R
\name{import_cats}
\alias{import_cats}
\title{Import cats data}
\usage{
import_cats(tagid)
}
\arguments{
\item{tagid}{A string with the tagid (e.g. "mn180105-22a"). Make sure the raw
data for the tag is on your hard drive!}
}
\value{
A tibble of raw tag data with fields:
\itemize{
  \item dateUTC (character)
  \item timeUTC (character)
  \item timeLocal (character)
  \item accX (double)
  \item accY (double)
  \item accZ (double)
  \item gyrX (double)
  \item gyrY (double)
  \item gyrZ (double)
  \item magX (double)
  \item magY (double)
  \item magZ (double)
  \item tempIMU (double)
  \item gpsDate (character)
  \item gpsTime (character)
  \item gps3 (integer)
  \item gpsSats (integer)
  \item depthM (double)
  \item depthDegC (double)
  \item light1 (integer)
  \item light2 (integer)
  \item syerr (character)
  \item battV (double)
  \item battmA (double)
  \item battmAh (double)
  \item camera (integer)
  \item flags (character)
  \item led (integer)
  \item camtime (integer)
  \item gpsL (logical)
  \item ccStatus (character)
  \item ccVidSz (integer)
}
}
\description{
\code{import_cats} reads and concatenates all the raw data CSVs from a CATS
tag. May take a long time and return a massive result (> 1GB) because it
hasn't been decimated yet.
}
\seealso{
\code{\link{list_raw_tags}}
}
