% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\name{init_prh}
\alias{init_prh}
\title{Initiate a PRH \code{init_prh} tries to load raw tag data and decimate it (by
default to 10 Hz).}
\usage{
init_prh(prh, new_freq)
}
\arguments{
\item{prh}{A PRH object.}

\item{new_freq}{The decimated sample frequency in Hz (by default 10 Hz).}
}
\value{
A PRH object with the following additional slots filled:
\itemize{
  \item freq
  \item rawdata
}
}
\description{
Initiating the PRH process requires importing the raw data and decimating it
to a manageable frequency. The next step is trimming to the tag on/off times.
}
\seealso{
\link{\code{trim_data}}, \link{\code{import_cats}}, \link{\code{decimate}}
}
