% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_rawdata.R
\name{interp_gaps}
\alias{interp_gaps}
\title{Linearly interpolate gaps}
\usage{
interp_gaps(prh, thr = 2)
}
\arguments{
\item{prh}{A PRH object.}

\item{thr}{A numeric defining the duration threshold (in sec) for maximum
interpolation window.}
}
\value{
A PRH object with the rawdata slot updated
}
\description{
\code{interp_gaps} interpolates gaps in sensor readings. Time skips and gaps
should be corrected first.
}
