% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_access.R
\name{list_raw_tags}
\alias{list_raw_tags}
\title{List all available deployments with raw data}
\usage{
list_raw_tags()
}
\value{
A string vector of all the tagids with available raw data.
}
\description{
List all available deployments with raw data
}
