% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_access.R
\name{load_cal}
\alias{load_cal}
\title{Load a calibration file}
\usage{
load_cal(tagnum)
}
\arguments{
\item{tagnum}{A string with the tag number (e.g. "44", not "mn180607-44").}
}
\value{
A list with fields: \itemize{ \item A \item acal \item aconst \item
  gycal \item gyconst \item magcaloff \item magcalon \item magconstoff \item
  magconston \item Tcal \item Tconst \item pcal \item pconst }
}
\description{
\code{load_cal} finds a tag's calibration .mat file and loads it into a list.
}
