% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\docType{class}
\name{PRH-class}
\alias{PRH-class}
\alias{.PRH}
\title{S4 class definition for PRH}
\description{
S4 class definition for PRH
}
\section{Slots}{

\describe{
\item{\code{tagid}}{A string with the tag ID e.g. "mn180607-44".}

\item{\code{tagnum}}{A string with the tag number e.g. "44"}

\item{\code{freq}}{A double with the sampling frequency after decimation in Hz
(typically 10 Hz).}

\item{\code{tagon,tagoff}}{A POSIXct with the date and time of tag attach and detach
in local time.}

\item{\code{slips}}{A POSIXct vector with the timestamps of each tag slip.}

\item{\code{W}}{A list of 3x3 numeric matrices to rotate tag frame into whale frame.
One matrix for each tag orientation (changes on slips).}

\item{\code{data}}{A tibble with the following columns: \itemize{ \item t A POSIXct
with the date and time of each record. \item p A numeric with the depth
records. \item rawA,rawM,rawG List columns with nx3 numeric matrices with
the raw accelerometer, magnetometer, and gyroscope records in engineering
units. \item At,Mt,Gt As rawA, rawM, and rawG, but calibrated in units of
g, microteslas, and radians per second, respectively in the tag's frame.
\item Aw,Mw,Gw As At, Mt, and Gt, but rotated into the whale's frame. \item
pitch,roll,head Numerics with the pitch, roll, and heading. \item
speedJJ,speedFN Numerics with speed estimates derived from jiggle and flow
noise, respectively.}}

\item{\code{rawdata}}{A tibble with the (decimated) raw data. Removed at the end of
the PRH creation process.}
}}

