% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_cats_path}
\alias{set_cats_path}
\title{Set path to CATS folder}
\usage{
set_cats_path(cats_path)
}
\arguments{
\item{cats_path}{A string with the path to the CATS folder.}
}
\description{
\code{set_cats_path} updates the config file with the location of the CATS
folder.
}
\details{
The CATS folder is typically on the COPYCATS hard drive. On a Mac the path is
usually /Volumes/COPYCATS/CATS and on Windows it's usually D:/CATS. The path
}
\examples{
\dontrun{
set_cats_path("/Volumes/COPYCATS/CATS")
}
}
