% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagonoff.R
\name{tagonoff}
\alias{tagonoff}
\title{Tag on/off GUI}
\usage{
tagonoff(rawdata)
}
\arguments{
\item{prh}{A tibble of raw tag data}
}
\value{
A PRH object with the tagon/off slots filled and the rawdata slot
  trimmed to the tagon/off times
}
\description{
\code{tagonoff} is a Shiny app that allows the user to zoom in on pressure
and acceleration to choose the tag on and off times. Should only be called
from \code{\link{trim_data}}.
}
