% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_access.R
\name{import_cats}
\alias{import_cats}
\title{Import cats data}
\usage{
import_cats(tagid)
}
\arguments{
\item{tagid}{A string with the tagid (e.g. "mn180105-22a"). Make sure the raw
data for the tag is on your hard drive!}
}
\value{
A tibble with the raw data from the tag.
}
\description{
\code{import_cats} reads and concatenates all the raw data CSVs from a CATS
tag. May take a long time and return a massive result (> 1GB) because it
hasn't been decimated yet.
}
\seealso{
\code{\link{list_raw_tags}}
}
