% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\name{init_prh}
\alias{init_prh}
\title{Initiate a PRH}
\usage{
init_prh(tagid, tagnum)
}
\arguments{
\item{tagid}{A string with the tagid (e.g. "mn180105-22a").}

\item{tagnum}{A string with the tag number (e.g. "22")}
}
\value{
A PRH object with the rawdata slot filled
}
\description{
\code{init_prh} tries to load raw tag data.
}
\details{
Initiating the PRH process finds the raw data on the CATS drive, imports it,
and tidies up the names.
}
\examples{
\dontrun{
prh_rawdata <- init_prh("mn180607-44", "44")
}
}
\seealso{
\code{\link{import_cats}} for reading raw data
}
