% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\docType{class}
\name{PRH-class}
\alias{PRH-class}
\alias{.PRH}
\title{S4 class definition for PRH}
\description{
S4 class definition for PRH
}
\section{Slots}{

\describe{
\item{\code{tagid}}{A string with the tag ID e.g. "mn180607-44".}

\item{\code{tagnum}}{A string with the tag number e.g. "44"}

\item{\code{freq}}{A double with the sampling frequency after decimation in Hz
(typically 10 Hz).}

\item{\code{tagon,tagoff}}{A POSIXct with the date and time of tag attach and detach
in local time.}

\item{\code{t}}{A POSIXct with the date and time of each record.}

\item{\code{rawA,rawM,rawG}}{An nx3 numeric matrix with the raw accelerometer,
magnetometer, and gyroscope records in enineering units.}

\item{\code{At,Mt,Gt}}{As rawA, rawM, and rawG, but calibrated in units of g,
microteslas, and radians per second, respectively in the tag's frame.}

\item{\code{Aw,Mw,Gw}}{As At, Mt, and Gt, but rotated into the whale's frame.}

\item{\code{W}}{A 3x3 numeric matrix to rotate tag frame into whale frame.}

\item{\code{pitch,roll,head}}{Numeric vectors with the pitch, roll, and heading.}

\item{\code{speedJJ,speedFN}}{Numeric vectors with speed estimates derived from
jiggle and flow noise, respectively.}
}}

