% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\name{PRH}
\alias{PRH}
\title{Constructor for PRH class}
\usage{
PRH(tagid, tagnum)
}
\arguments{
\item{tagid}{A string with the tagid (e.g. "mn180105-22a").}
}
\value{
An object of class PRH with the tagid and tagnum slots filled.
}
\description{
Constructor for PRH class
}
