% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagonoff.R
\name{tagonoff}
\alias{tagonoff}
\title{Tag on/off GUI}
\usage{
tagonoff(rawdata)
}
\arguments{
\item{rawdata}{A tibble of raw tag data (e.g. returned from
\code{import_cats})}
}
\description{
\code{tagonoff} is a Shiny app that allows the user to zoom in on pressure
and acceleration to choose the tag on and off times.
}
